# internal/infrastructure/entrypoints/apiRest/routes/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Service to connect users",
        "title": "Bombink End Points",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:1323",
    "basePath": "/commerce/manage-bombink",
    "paths": {
        "/commerce/manage-bombink/user/create": {
            "post": {
                "description": "Service to activate missions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ActivateUser"
                ],
                "summary": "Service to activate missions",
                "parameters": [
                    {
                        "description": "Save user",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.BombinkResponse"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "user guardado con éxito",
                        "schema": {
                            "$ref": "#/definitions/dtos.BombinkResponse"
                        }
                    },
                    "500": {
                        "description": "Cuando ocurre un error inexperado",
                        "schema": {
                            "$ref": "#/definitions/api.RetomaRegistroResponseErrorInexperado"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Data": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "IDMissionNoValid"
                }
            }
        },
        "api.RetomaRegistroResponseErrorInexperado": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.Data"
                },
                "message": {
                    "type": "string",
                    "example": "Ocurrio un error inexperado"
                }
            }
        },
        "dtos.BombinkResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}